Add support to load and render skeletons of MSR Action 3D

+++ KinectProject/types/DepthFrame.h
@@ -19,8 +19,8 @@ class DepthFrame : public DataFrame
 public:
     // Static Class Methods
     static void calculateHistogram(QMap<float, float> &pHistogram, const DepthFrame &frame);
-    static int maxValue(const DepthFrame& frame);
-    static int minValue(const DepthFrame& frame);
+    static float maxValue(const DepthFrame& frame);
+    static float minValue(const DepthFrame& frame);
 
     // Constructor, Destructors and Copy Constructor
     explicit DepthFrame();
